requirements.txt
APScheduler==3.3.1
boto3==1.4.4
botocore==1.5.71
certifi==2017.4.17
chardet==3.0.4
docutils==0.13.1
future==0.16.0
idna==2.5
jmespath==0.9.3
oauthlib==2.0.2
python-dateutil==2.6.0
python-twitter==3.3
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
tzlocal==1.4
urllib3==1.21.1
